Update home.html
Edited this template so the user lands on this as the main page and see the basic info of the app so far. User can go from here to how it works, to see the basic info of how the page works so far (if it is a new user). Also, the user can go to login/signup page and login to their user and redirects to their dashboard (by using a form), also if he wants to create an account, it will redirects to a form and put his information