Updated views.py in landing
Deleted all stuff (packages, etc) related to django default sql server and now the code is directly focus on our database in mongodb.

Made the user_id field to be automatically created, so a new user will be assigned the next available number in the list.

date of creation is automatically captured using datetime.now

still need to check the dashboard part in this file, to ensure it works correctly